refactor/component 구분
