+++ jsSecond/Mission5/style.css
@@ -73,5 +73,5 @@ body {
 .scroll-observer {
   background-color: red;
   text-align: center;
-  margin: 10px;
+  margin: 0px;
 }
